tests/testrequirements.txt
# tracker requirements
boto3==1.20.31
fsspec==2021.11.1
python-dateutil==2.8.2
pytest==6.2.5
pytest-cases==3.6.5
pytest-env==0.6.2
pytest-mock==3.6.1
pytest-pythonpath==0.7.3
rstr==3.0.0
moto==2.2.19
pandas==1.3.4
# orjson==3.6.5
orjson
# svix-ksuid==0.6.0
svix-ksuid
fastparquet==0.7.2
# this works
s3fs==0.4.2
# new version of s3fs does not work in tests
# s3fs==2021.11.1
portion==2.2.0
zstandard==0.16.0

# trainer requirements
coremltools  # ~=5.2
# this is only temporary solution (version 3.20+ breaks protobuf):
protobuf~=3.19.0
# consider pinning xgb version if improveai's xgboost version will be bumped to most recent
# xgboost==1.4.2
dask
dask_ml
# coremltools limits max usable version to 1.4.2 -> check once in  a while
xgboost==1.4.2
xxhash
